Bug 1847469 - Part 15: Use uint32_t for line and column in debugger as a preparation to use column number types. r=iain
Differential Revision: https://phabricator.services.mozilla.com/D185753

UltraBlame original commit: 5df4f43ac0bfea7021e39462fa90879d79c98922

+++ js/src/wasm/WasmDebug.h
@@ -106,8 +106,8 @@ class DebugState {
 
   [[nodiscard]] bool getLineOffsets(size_t lineno, Vector<uint32_t>* offsets);
   [[nodiscard]] bool getAllColumnOffsets(Vector<ExprLoc>* offsets);
-  [[nodiscard]] bool getOffsetLocation(uint32_t offset, size_t* lineno,
-                                       size_t* column);
+  [[nodiscard]] bool getOffsetLocation(uint32_t offset, uint32_t* lineno,
+                                       uint32_t* column);
